MainCharacter Script and Animation Update
Started working on the MainCharacter script. We now can track his on screen position, move the player around using the on-screen joystick, and the animation plays in any given direction. I still need to improve the state machine as there seems to be a 1-2 second delay in starting the animations. I just wanted the group to have a rough idea before 3:30 Tuesday. This was all done in the CharacterMovementTest Scene.

Thank you.