Added TextBox.cs and Instantiate.cs
When you get close enough to a Friendly NPC, represented as a bookshelf in this instance, a pop-up box will appear. Will adapt for more objects. This is needed to decide whether to kill or save a friendly, thus increasing or decreasing the sanity. Code is commented; please let me know if clarification is needed.